timer is now Long instead of int

+++ src/main/java/ch/uzh/ifi/hase/soprafs22/rest/dto/LobbyPostDTO.java
@@ -9,7 +9,7 @@ public class LobbyPostDTO {
 
     private Long total_players;
 
-    private int timer;
+    private Long timer;
 
     public Long getHost_id() {
         return host_id;
@@ -43,7 +43,7 @@ public class LobbyPostDTO {
         this.total_players = total_players;
     }
 
-    public int getTimer() { return timer; }
+    public Long getTimer() { return timer; }
 
-    public void setTimer(int timer) { this.timer = timer; }
+    public void setTimer(Long timer) { this.timer = timer; }
 }
